copy: add fast copy path for darwin
